make: build static libXNVCtrl.a in tree
Download the upstream nvidia-settings source code, which contains the
libXNVCtrl source code, for the driver version set in driver_version.mk.
Then build libXNVCtrl statically and use the existing build hooks to
pass in the path automatically.

I don't know of a way to access the driver version dynamically that
doesn't rely on this library, which is a chicken-egg problem, or the
package manager, whose information is not portable.

I spent way too much time on Ubuntu finding out that the library was in
/usr/lib/x86_64-linux-gnu, which src/nvctrl/build.rs doesn't search by
default. I then spent even more time on openSUSE Leap 15.2 finding out
that its library path is /usr/lib64, which build.rs also doesn't search
by default, and that it only provides the dynamic library because
"static is bad, m'kay", and then not even dynamic linking seemed to
work [1] because X hadn't granted permission to control the fans [2].
When nvfancontrol finally had permission I had already written all of
this supporting infrastructure and now perhaps I don't have to spend
time finding the library in package managers and trees any more...

[1] The actual linking worked fine.
[2] Commit 6646466 ("Add basic Makefile as convenient Cargo proxy",
    2020-12-17)