Track Cargo.lock
The test test_invalid_toml_from_string fails because of an upstream
change in toml:0.5.7, which snuck in because Cargo.lock wasn't tracked.
This is basically the same problem as the one discussed in commit
bcfd45d ("Integrate GitHub Actions", 2020-04-14).

Let Cargo.lock files be tracked, and track old versions that build and
pass tests.

There is never any reason not to track Cargo.lock files. Cargo ignores
them for libraries and users want them for applications (to avoid this
error silly).